localtestazureblob/requirements.txt
mlflow==2.8.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
